issue #7 変更: WeatherCondition クラスを別ファイルに移す
