+++ lib/domain/models/weather.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/foundation.dart';
-import 'package:flutter_training/core/exceptions/weather_exceptions.dart';
+import 'package:flutter_training/domain/models/weather_condition.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
 part 'weather.freezed.dart';
@@ -11,19 +11,4 @@ class Weather with _$Weather {
     required int maxTemperature,
     required int minTemperature,
   }) = _Weather;
-}
-
-enum WeatherCondition {
-  sunny,
-  cloudy,
-  rainy;
-
-  factory WeatherCondition.from(String condition) {
-    for (final value in WeatherCondition.values) {
-      if (value.name == condition) {
-        return value;
-      }
-    }
-    throw UndefinedWeatherException(message: 'Undefined weather: $condition');
-  }
 }
